requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
tensorflow-cpu
numpy